[Feat] config server 적용

+++ src/main/java/com/example/userservice/dto/ContributionRes.java
@@ -8,6 +8,6 @@ public class ContributionRes {
     private Integer studentNumber;
     private String name;
     private String pictureUrl;
-    private Long contribution;
+    private Float contribution;
 
 }
